Websafe vC1.0

+++ insecure_websafe/comment/delete_comment.php
@@ -3,7 +3,10 @@
 if (isset($_GET["com_id"])) {
     $comment_id = $_GET["com_id"];
    
-    include "../sql_con.php";
+    $con = mysqli_connect("database","Lottie","Ad0r@ble","websafe");
+    if (!$con) {
+        die("Failed to connect " .  mysqli_connect_errno());
+    }
 
     $query = $con->prepare('DELETE FROM `comments` WHERE `comment_id` = ?');
     $query->bind_param("i", $comment_id);
